+++ Magic/CHANGELOG.md
@@ -4,6 +4,8 @@
 
  - Add left_click_sneak, right_click_sneak, drop_sneak and swap_sneak wand actions
  - The "examples" list in config.yml can now be put on a single line
+ - Fix log spam when swinging a wand that has no active spell
+ - Fix mana not saving properly on droppable wands
 
 # 7.9.6