+++ Magic/CHANGELOG.md
@@ -7,6 +7,7 @@
  - Fix log spam when swinging a wand that has no active spell
  - Fix mana not saving properly on droppable wands
  - Fix Colloportus in 1.14+
+ - Add require_mana_percentage parameter to CheckMana action
 
 # 7.9.6