+++ Magic/CHANGELOG.md
@@ -11,6 +11,7 @@
  - Add require_mana_percentage parameter to CheckMana action
  - Add item_count parameter to GiveItem action so it can be used with variables
  - Fix dupe bug/exploit with the swap key
+ - Fix simple potion_effect maps
 
 # 7.9.6