+++ Magic/CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+# 7.9.8
+
 # 7.9.7
 
  - Moved engineering magic into a separate example